telemetry

+++ MIDAS/src/hilsim/telemetry_backend.h
@@ -12,9 +12,9 @@ public:
     explicit TelemetryBackend(const char* file_name);
     ErrorCode __attribute__((warn_unused_result)) init();
 
-    int8_t getRecentRssi();
+    int16_t getRecentRssi();
     void setFrequency(float frequency);
-
+    
     template<typename T>
     void send(const T& data) {
         output_file.write((const char*) &data, sizeof(T));
